Make compose_sprite() consistent under Perl 5.18 hash key ordering
In Perl 5.18 the hash key order randomization was made much more random
much more often. For CSS::SpriteMaker this caused the order of composed
sprites to be random, whereas they should be in the order of the parts
arrayref argument.

Solution was basically to added a bunch of sort() calls in front of all
the calls to keys(), to ensure completely consistent ordering.